Move configuration data to a separate file.
Instead of asking the user to edit updater.sh, move the configuration
data to a separate config file.  Optionally, if the user wants to put
the configuration in ~/.config/googleDNS, then the user won't have to
modify a file tracked by git.

While we're at it, set up things so that the response.log file in the
googleDNS directory will be written (as opposed to whatever the
current directory is when updater.sh is run).

Signed-off-by: Theodore Ts'o <tytso@mit.edu>